Use floats, use interpolation to reduce BC calls

+++ src/3dUtils.h
@@ -2,45 +2,45 @@
 #include "types.h"
 
 // Calculates the distance from a point to a line
-double DistanceToLine(Vec3d const &line_point1, Vec3d const &line_point2) {
-    Vec3d line = line_point2 - line_point1;
-    double t = line_point1.dot(line) / line.norm();
+float DistanceToLine(Vec3f const &line_point1, Vec3f const &line_point2) {
+    Vec3f line = line_point2 - line_point1;
+    float t = line_point1.dot(line) / line.norm();
     if (t < 0) {
         return line_point1.norm();
     } else if (t > 1) {
         return line_point2.norm();
     } else {
-        Vec3d projection = line_point1 + t * line;
+        Vec3f projection = line_point1 + t * line;
         return projection.norm();
     }
 }
 
-Vec3d Normal(Vec3d const &p1, Vec3d const &p2, Vec3d const &p3) {
-    Vec3d cross = (p2 - p1).cross(p3 - p1);
+Vec3f Normal(Vec3f const &p1, Vec3f const &p2, Vec3f const &p3) {
+    Vec3f cross = (p2 - p1).cross(p3 - p1);
     return cross.normalized();
 }
 
-bool FacesCamera(Eigen::Matrix<double, 3, 3> const& points) {
+bool FacesCamera(Eigen::Matrix<float, 3, 3> const &points) {
     // Calculate the normal of the triangle
-    Vec3d normal = Normal(points.col(0), points.col(1), points.col(2));
-    double d = normal.dot(points.col(0));
+    Vec3f normal = Normal(points.col(0), points.col(1), points.col(2));
+    float d = normal.dot(points.col(0));
     return d > 0;
 }
 
-double NormToPoint(Eigen::Matrix<double, 3, 3> const& points, Vec3d const &p) {
+float NormToPoint(Eigen::Matrix<float, 3, 3> const &points, Vec3f const &p) {
     // Calculate the normal of the triangle
-    Vec3d normal = Normal(points.col(0), points.col(1), points.col(2));
-    Vec3d vecToP = (p - points.col(0)).normalized();
-    double facePoint = -normal.dot(vecToP);
+    Vec3f normal = Normal(points.col(0), points.col(1), points.col(2));
+    Vec3f vecToP = (p - points.col(0)).normalized();
+    float facePoint = -normal.dot(vecToP);
     return facePoint;
 }
 
-double ShortestDistance(Vec3d const &t1, Vec3d const &t2, Vec3d const &t3) {
+float ShortestDistance(Vec3f const &t1, Vec3f const &t2, Vec3f const &t3) {
     // Calculate the normal of the triangle
-    Vec3d normal = Normal(t1, t2, t3);
+    Vec3f normal = Normal(t1, t2, t3);
 
     // Calculate the distance from the point to the plane of the triangle
-    double d = normal.dot(t1);
+    float d = normal.dot(t1);
 
     // If the point is on the same side of the plane as the normal, the shortest
     // distance is the distance from the point to the plane
@@ -52,4 +52,3 @@ double ShortestDistance(Vec3d const &t1, Vec3d const &t2, Vec3d const &t3) {
     return std::min(
         {DistanceToLine(t1, t2), DistanceToLine(t2, t3), DistanceToLine(t3, t1), t1.norm(), t2.norm(), t3.norm()});
 }
-
